Update Final code demo 2

+++ SourceCode/17520897-17520407/CPanther.h
@@ -34,8 +34,8 @@ private:
 	float xGround;
 	float yGround;
 
-	int widthGround;
-	int heightGround;
+	float widthGround;
+	float heightGround;
 
 	DWORD timeSpawn;
